+++ bindings/matlab/+iDynTreeWrappers/prepareVisualization.m
@@ -13,7 +13,10 @@ function [Visualizer,Objects]=prepareVisualization(KinDynModel,meshFilePrefix,va
 %     - `groundTransparency` : Selects the transparency of the ground.
 %     - `groundFrame` : Selects the frame in which the ground is attached.
 %     - `name` : The name of the figure 
-%     - `reuseFigure` : Enable the reuse of an already open figure with the same name
+%     - `reuseFigure` : Enable the reuse of an already open figure. It can be the following values:
+%         - true: Reuse the figure with the same name (the figure is cleared before reusing it)
+%         - 'gcf': Reuse the figure returned by gcf
+%         - false: Do not reuse the figure.
 %     Note: all extra variables are sent to `plotMeshInWorld`
 %   - Outputs:
 %       - `Visualizer` : Struct containing the following fields
@@ -59,7 +62,7 @@ addParameter(p,'groundColor',default_groundColor,@(x)validateattributes(x,{'nume
 addParameter(p,'groundTransparency',default_groundTransparency,@(x) isnumeric(x) && isscalar(x));
 addParameter(p,'groundFrame',default_groundFrame,@(x) isstring(x) || ischar(x));
 addParameter(p,'name',default_name,@(x) isstring(x) || ischar(x));
-addParameter(p,'reuseFigure',default_reuseFigure,@(x) islogical(x));
+addParameter(p,'reuseFigure',default_reuseFigure,@(x) islogical(x) || strcmp(x,'gcf'));
 
 % parse inputs
 parse(p,KinDynModel,meshFilePrefix,varargin{:});
@@ -70,7 +73,9 @@ model=KinDynModel.kinDynComp.model;
 numberOfLinks=length(linkMeshInfo);
 linkNames=cell(numberOfLinks,1);
 figHandles = findobj('Type', 'figure', 'Name', options.name);
-if options.reuseFigure && size(figHandles, 1) > 0
+if strcmp(options.reuseFigure,'gcf')
+    mainHandler=gcf;
+elseif options.reuseFigure && size(figHandles, 1) > 0
     mainHandler=figHandles(1,1);
     clf(mainHandler,'reset')
 else